requirements.txt
python-dotenv==1.0.1
Flask==3.0.2
Flask-PyMongo==3.0.1
requests==2.32.3
flask_cors==5.0.1